+++ test/unit/fuzz-tools.test.js
@@ -115,6 +115,7 @@ import * as fluidrespV113 from '../../lib/fluidresp-v113.js';
 import * as pulmV114 from '../../lib/pulm-v114.js';
 import * as pulmnodV115 from '../../lib/pulmnod-v115.js';
 import * as neuroV117 from '../../lib/neuro-v117.js';
+import * as neuroV118 from '../../lib/neuro-v118.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -174,6 +175,7 @@ const MODULES = {
   'pulm-v114.js': pulmV114,
   'pulmnod-v115.js': pulmnodV115,
   'neuro-v117.js': neuroV117,
+  'neuro-v118.js': neuroV118,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];